So I fixed Judauism.
So the character data for Judea did not work since the New character Ids did not want to work with the old ones. Also to have a character you must have a "living" tag, and since Juda was a dead tag, I moved the characters to HMJ. I also added constantine to jesus heaven, and fixed zoroaster not showing up.